feat(releases): :rocket: Created file to execute DB release tasks when run in production
We want our application to run database migrations when the application starts up. Locally in dev, we use mix ecto.migrate. In production when using an Elixir Release, we don't have the mix tool available to help out. Let's create a file called lib/hello_elixir/release.ex to help with this